Alteração de atributo de double para BigDecimal

+++ src/main/java/com/pdv/lalapan/dto/ProdutoEstoqueBaixoDTO.java
@@ -2,7 +2,9 @@ package com.pdv.lalapan.dto;
 
 import com.pdv.lalapan.entities.Produto;
 
-public record ProdutoEstoqueBaixoDTO(Long id, String nome, double quantidadeEstoque, double estoqueMinimo) {
+import java.math.BigDecimal;
+
+public record ProdutoEstoqueBaixoDTO(Long id, String nome, BigDecimal quantidadeEstoque, BigDecimal estoqueMinimo) {
 
     public static ProdutoEstoqueBaixoDTO fromEntity(Produto p) {
         return new ProdutoEstoqueBaixoDTO(
