+++ src/main/java/com/pdv/lalapan/handler/DetalhesEstoque.java
@@ -2,5 +2,5 @@ package com.pdv.lalapan.handler;
 
 import java.math.BigDecimal;
 
-public record DetalhesEstoque(String nomeProduto, BigDecimal disponivel, Double solicitado) {
+public record DetalhesEstoque(String nomeProduto, BigDecimal disponivel, BigDecimal solicitado) {
 }
